gcf/admin/users/requirements.txt
firebase-admin
Flask>=1.0,<3.0
functions-framework==3.0.0
google-cloud-firestore==2.9.1
